requirements.txt
jina
pandas
numpy
transformers==4.12.3
newspaper3k
sentencepiece
torch